Handle missing attribute files when parsing
Previously using Gitlab::Git::Attributes when $GIT_DIR/info/attributes
didn't exist would result in a runtime error. This commit fixes this so
an empty Hash is produced instead.

+++ lib/gitlab_git/attributes.rb
@@ -86,6 +86,8 @@ module Gitlab
       def each_line
         full_path = File.join(@path, 'info/attributes')
 
+        return unless File.exist?(full_path)
+
         File.open(full_path, 'r') do |handle|
           handle.each_line do |line|
             yield line.strip
